- Splitting out WidePos from Pos.h  - Adding a Rect class for dimensions  - Adding a make_unique.hpp for compilers that don't support C++14 yet with an implementation from the N3656 paper  - Making it so that GMs don't have to complete the Time Temple quests to proceed  - Changing most places that reference GM hide and GM equips separately to using a single bool function  - Adding "hhg" to !map teleport options  - Converting a vast array of iterator loops to range-based for loops  - Converting most of the naked new calls with shared_ptr and unique_ptr that I could find to make_shared or make_unique respectively (there are some I need to evaluate specially and some I probably just plain missed - I also need to reevaluate the vast quantity of regular naked pointers)  - Refactoring log replacements so the same brick of code isn't found in 3 separate spots  - Modifying many map lookups to utilize the iterator instead of doing multiple lookups (need to do this for many, many more)  - Removing pointers from KeyMaps  - Consolidating all map timer operations into a single timer  - Changing SeatInfo to carry an ID  - Making some of the Map init functions private + adding MapDataProvider as a friend  - Changing many of the APIs and data structures to use the new Rect class to represent lt/rb  - Changing the name of giveExp in Mob to getHighestDamager to reflect the value it returns (although this also distributes EXP.. will consider revising the structure of how this works)  - Adding repeating EXP messages in the case where EXP is > int32_t's max value (so 2147483747 will show two messages: EXP +2147483647 and EXP +100)  - Beginning to look into getting a legitimate spawning algorithm  - As usual, I probably broke some stuff and I can only guarantee it compiles for MSVC right now
git-svn-id: https://dev.chiasoft.net:8443/svn/svn_vana/branches/buis_vana_overhaul@3284 23f3cdc0-e6d1-42be-a0c1-51f4125517bb

+++ src/ChannelServer/PlayerStats.h
@@ -96,7 +96,7 @@ public:
 	void setEquip(int16_t slot, Item *equip, bool isLoading = false);
 
 	// Level related functions
-	void giveExp(uint32_t exp, bool inChat = false, bool white = true);
+	void giveExp(uint64_t exp, bool inChat = false, bool white = true);
 	void addStat(PacketReader &packet);
 	void addStatMulti(PacketReader &packet);
 	void addStat(int32_t type, int16_t mod = 1, bool isReset = false);
